+++ forui/lib/src/widgets/popover_menu.dart
@@ -17,7 +17,15 @@ part 'popover_menu.style.dart';
 /// * [FPopoverController] for controlling a popover menu.
 /// * [FPopoverMenuStyle] for customizing a popover menu's appearance.
 /// * [FTileGroup] for customizing the items in the menu.
-class FPopoverMenu extends StatefulWidget {
+class FPopoverMenu extends StatelessWidget {
+  static List<FTileGroupMixin<FTileMixin>> _menuBuilder(
+    BuildContext context,
+    FPopoverController controller,
+    List<FTileGroupMixin<FTileMixin>>? menu,
+  ) => menu!;
+
+  static Widget _builder(BuildContext _, FPopoverController _, Widget? child) => child!;
+
   /// The popover menu's style.
   ///
   /// ## CLI
@@ -79,6 +87,9 @@ class FPopoverMenu extends StatefulWidget {
   /// {@macro forui.widgets.FPopover.hideOnTapOutside}
   final FHidePopoverRegion hideOnTapOutside;
 
+  /// {@macro forui.widgets.FPopover.barrier}
+  final ImageFilter? barrier;
+
   /// {@macro forui.foundation.doc_templates.autofocus}
   final bool autofocus;
 
@@ -94,19 +105,35 @@ class FPopoverMenu extends StatefulWidget {
   /// The menu's semantic label used by accessibility frameworks.
   final String? semanticsLabel;
 
+  /// An optional builder which returns the menu that the popover is aligned to.
+  ///
+  /// Can incorporate a value-independent widget subtree from the [menu] into the returned widget tree.
+  ///
+  /// This can be null if the entire widget subtree the [menuBuilder] builds doest not require the controller.
+  final List<FTileGroupMixin<FTileMixin>> Function(BuildContext, FPopoverController, List<FTileGroupMixin<FTileMixin>>?)
+  menuBuilder;
+
   /// The menu.
-  final List<FTileGroupMixin<FTileMixin>> menu;
+  ///
+  /// Passed to [menuBuilder] if provided.
+  final List<FTileGroupMixin<FTileMixin>>? menu;
+
+  /// {@macro forui.widgets.FPopover.builder}
+  final ValueWidgetBuilder<FPopoverController> builder;
 
   /// The child.
-  final Widget child;
-
-  final bool _automatic;
+  ///
+  /// Passed to [builder] if provided.
+  final Widget? child;
 
   /// Creates a menu that only shows the menu when the controller is manually toggled.
+  ///
+  /// ## Contract
+  /// Throws [AssertionError] if:
+  /// * both [builder] and [child] are provided.
+  /// * both [menuBuilder] and [menu] are null.
   const FPopoverMenu({
-    required FPopoverController this.popoverController,
-    required this.menu,
-    required this.child,
+    this.popoverController,
     this.scrollController,
     this.style,
     this.cacheExtent,
@@ -120,46 +147,53 @@ class FPopoverMenu extends StatefulWidget {
     this.offset = Offset.zero,
     this.groupId,
     this.hideOnTapOutside = FHidePopoverRegion.anywhere,
+    this.barrier,
     this.semanticsLabel,
     this.autofocus = false,
     this.focusNode,
     this.onFocusChange,
     this.traversalEdgeBehavior = TraversalEdgeBehavior.closedLoop,
+    this.menuBuilder = _menuBuilder,
+    this.menu,
+    this.builder = _builder,
+    this.child,
     super.key,
-  }) : _automatic = false;
-
-  /// Creates a menu that is automatically shown when the [child] is tapped.
-  ///
-  /// It is not recommended for the [child] to contain a [GestureDetector], such as [FButton]. Only one
-  /// `GestureDetector` will be called if there are multiple overlapping `GestureDetector`s, leading to unexpected
-  /// behavior.
-  const FPopoverMenu.automatic({
-    required this.menu,
-    required this.child,
-    this.style,
-    this.popoverController,
-    this.scrollController,
-    this.cacheExtent,
-    this.maxHeight = double.infinity,
-    this.dragStartBehavior = DragStartBehavior.start,
-    this.divider = FTileDivider.full,
-    this.menuAnchor = Alignment.topCenter,
-    this.childAnchor = Alignment.bottomCenter,
-    this.spacing = const FPortalSpacing(4),
-    this.shift = FPortalShift.flip,
-    this.offset = Offset.zero,
-    this.groupId,
-    this.hideOnTapOutside = FHidePopoverRegion.excludeTarget,
-    this.semanticsLabel,
-    this.autofocus = false,
-    this.focusNode,
-    this.onFocusChange,
-    this.traversalEdgeBehavior = TraversalEdgeBehavior.closedLoop,
-    super.key,
-  }) : _automatic = true;
+  }) : assert(builder != _builder || child != null, 'Either builder or child must be provided.'),
+       assert(menuBuilder != _menuBuilder || menu != null, 'Either menuBuilder or menu must be provided.');
 
   @override
-  State<FPopoverMenu> createState() => _FPopoverMenuState();
+  Widget build(BuildContext context) {
+    final style = this.style ?? context.theme.popoverMenuStyle;
+    return FPopover(
+      controller: popoverController,
+      style: style,
+      constraints: FPortalConstraints(maxWidth: style.maxWidth),
+      popoverAnchor: menuAnchor,
+      childAnchor: childAnchor,
+      spacing: spacing,
+      shift: shift,
+      offset: offset,
+      groupId: groupId,
+      hideOnTapOutside: hideOnTapOutside,
+      barrier: barrier,
+      autofocus: autofocus,
+      focusNode: focusNode,
+      onFocusChange: onFocusChange,
+      traversalEdgeBehavior: traversalEdgeBehavior,
+      popoverBuilder: (context, controller) => FTileGroup.merge(
+        scrollController: scrollController,
+        cacheExtent: cacheExtent,
+        maxHeight: maxHeight,
+        dragStartBehavior: dragStartBehavior,
+        semanticsLabel: semanticsLabel,
+        style: style.tileGroupStyle,
+        divider: divider,
+        children: menuBuilder(context, controller, menu),
+      ),
+      builder: builder,
+      child: child,
+    );
+  }
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -179,66 +213,14 @@ class FPopoverMenu extends StatefulWidget {
       ..add(DiagnosticsProperty('offset', offset))
       ..add(DiagnosticsProperty('groupId', groupId))
       ..add(EnumProperty('hideOnTapOutside', hideOnTapOutside))
+      ..add(DiagnosticsProperty('barrier', barrier))
       ..add(StringProperty('semanticsLabel', semanticsLabel))
       ..add(FlagProperty('autofocus', value: autofocus, ifTrue: 'autofocus'))
       ..add(DiagnosticsProperty('focusNode', focusNode))
       ..add(ObjectFlagProperty.has('onFocusChange', onFocusChange))
-      ..add(EnumProperty('traversalEdgeBehavior', traversalEdgeBehavior));
-  }
-}
-
-class _FPopoverMenuState extends State<FPopoverMenu> with SingleTickerProviderStateMixin {
-  late FPopoverController _popoverController = widget.popoverController ?? FPopoverController(vsync: this);
-
-  @override
-  void didUpdateWidget(covariant FPopoverMenu old) {
-    super.didUpdateWidget(old);
-    if (widget.popoverController != old.popoverController) {
-      if (old.popoverController == null) {
-        _popoverController.dispose();
-      }
-      _popoverController = widget.popoverController ?? FPopoverController(vsync: this);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final style = widget.style ?? context.theme.popoverMenuStyle;
-    return (widget._automatic ? FPopover.automatic : FPopover.new)(
-      controller: _popoverController,
-      style: style,
-      constraints: FPortalConstraints(maxWidth: style.maxWidth),
-      popoverAnchor: widget.menuAnchor,
-      childAnchor: widget.childAnchor,
-      spacing: widget.spacing,
-      shift: widget.shift,
-      offset: widget.offset,
-      groupId: widget.groupId,
-      hideOnTapOutside: widget.hideOnTapOutside,
-      autofocus: widget.autofocus,
-      focusNode: widget.focusNode,
-      onFocusChange: widget.onFocusChange,
-      traversalEdgeBehavior: widget.traversalEdgeBehavior,
-      popoverBuilder: (_, _, _) => FTileGroup.merge(
-        scrollController: widget.scrollController,
-        cacheExtent: widget.cacheExtent,
-        maxHeight: widget.maxHeight,
-        dragStartBehavior: widget.dragStartBehavior,
-        semanticsLabel: widget.semanticsLabel,
-        style: style.tileGroupStyle,
-        divider: widget.divider,
-        children: widget.menu,
-      ),
-      child: widget.child,
-    );
-  }
-
-  @override
-  void dispose() {
-    if (widget.popoverController == null) {
-      _popoverController.dispose();
-    }
-    super.dispose();
+      ..add(EnumProperty('traversalEdgeBehavior', traversalEdgeBehavior))
+      ..add(ObjectFlagProperty.has('menuBuilder', menuBuilder))
+      ..add(ObjectFlagProperty.has('builder', builder));
   }
 }